added commetns in backend files
